pyproject.toml
[project]
name = "chatmachine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "tomaszbk", email = "zubik.tomas@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "discord>=2.3.2",
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
